fix: @putout/operator-filesystem: removeEmptyDirectory: root

+++ packages/operator-filesystem/lib/filesystem.js
@@ -193,6 +193,10 @@ module.exports.removeEmptyDirectory = (dirPath) => {
             break;
         
         nextParentDir = getParentDirectory(dirPath);
+        
+        if (!nextParentDir)
+            break;
+        
         removeFile(dirPath);
         dirPath = nextParentDir;
     }
